+++ src/components/UI/LocaleSwitcher.tsx
@@ -1,24 +1,48 @@
-import { useLocale, useTranslations } from "next-intl";
-import LocaleSwitcherSelect from "./LocaleSwitcherSelect";
+"use client";
+import { Locale, useLocale, useTranslations } from "next-intl";
+import { useTransition } from "react";
+import { setUserLocale } from "@/services/locale";
+import { LanguageSelector } from "@trussworks/react-uswds";
 
 export default function LocaleSwitcher() {
   const t = useTranslations("LocaleSwitcher");
   const locale = useLocale();
+  const [isPending, startTransition] = useTransition();
+
+  // TODO automate this from the locales in the config
+  // TODO use label_local somehow
+  // https://trussworks.github.io/react-uswds/?path=/docs/components-languageselector--docs
+  const items: { value: Locale; label: string }[] = [
+    {
+      value: "en",
+      label: t("en"),
+    },
+    {
+      value: "es",
+      label: t("es"),
+    },
+  ];
+
+  function onChange(loc) {
+    startTransition(() => {
+      setUserLocale(loc);
+    });
+  }
+
+  const langs = items.map((item) => ({
+    attr: item.value,
+    label: item.label,
+    on_click: () => {
+      onChange(item.value);
+    },
+  }));
 
   return (
-    <LocaleSwitcherSelect
+    <LanguageSelector
+      id="locale-select"
+      langs={langs}
+      className={isPending ?? "opacity-50"}
       defaultValue={locale}
-      items={[
-        {
-          value: "en",
-          label: t("en"),
-        },
-        {
-          value: "es",
-          label: t("es"),
-        },
-      ]}
-      label={t("label")}
     />
   );
 }
